+++ src/main/scala/org/podval/tools/scalajs/jvm/JvmDelegate.scala
@@ -1,8 +1,7 @@
 package org.podval.tools.scalajs.jvm
 
 import org.gradle.api.Project
-import org.gradle.api.plugins.JavaPlugin
-import org.podval.tools.build.{JavaDependency, ScalaLibrary, Version}
+import org.podval.tools.build.ScalaLibrary
 import org.podval.tools.scalajs.ScalaJSPlugin
 
 class JvmDelegate extends ScalaJSPlugin.Delegate:
@@ -14,12 +13,4 @@ class JvmDelegate extends ScalaJSPlugin.Delegate:
     pluginScalaLibrary : ScalaLibrary,
     projectScalaLibrary: ScalaLibrary
   ): Unit =
-    JavaDependency.Requirement(
-      dependency = JavaDependency(group = "org.scala-sbt", artifact = "test-interface"),
-      version = Version("1.0"),
-      reason =
-        """because some test frameworks (ScalaTest :)) do not bring it in in,
-          |and it needs to be on the testImplementation classpath
-          |""".stripMargin,
-      configurationName = JavaPlugin.TEST_IMPLEMENTATION_CONFIGURATION_NAME
-    ).applyToConfiguration(project)
+    JvmDependencies.sbtTestInterfaceDependencyRequirement.applyToConfiguration(project)