+++ CHANGELOG.md
@@ -23,6 +23,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - cleanup: Node setup;
 - cleanup: split the plugin class into `ScalaDelegate` and `ScalaJSDelegate`;
 - cleanup: Gradle;
+- doc: converted `README` from MarkDown to AsciiDoc;
+- doc: refreshed `README`;
+- doc: write the AsciiDoc version attributes into a separate file programmatically and include it in `README`;
+- build: set dependency versions from properties in `gradle.properties` written programmatically;
 
 ## [0.4.15] - 2025-01-17
 - cleanup: minor cleanup